requirements.txt
PyQt6>=6.4.0
opencv-python>=4.8.0
face_recognition>=1.3.0
face_recognition_models>=0.3.0
numpy>=1.24.0
Pillow>=10.0.0
dlib>=19.24.0
cmake>=3.25.0